requirements.txt
fastapi[standard]==0.115.8
uvicorn==0.34.0
pydantic==2.10.6

Pillow==11.1.0
inky[rpi]==2.0.0
torchaudio==2.6.0
torchvision==0.21.0
openai==1.62.0
soundfile==0.13.1

common @ git+https://git@github.com/aureka-team/common.git@v1.1.2
